general LWRP improvements
This LWRP, for the most part, solves all of my needs. There are a few
assumptions that it makes which I think should be left up to the consumer. The
biggest concern is that it's restarting Nginx on a configuration change, which
might not be preferred by some (with reload being preferred). We should remove
the notification of `service[nginx]` from the LWRP and leave it up to the
consumer to notify.

Unfortunately, that's not possible right now. Currently, the `nginx_conf_file`
LWRP is not seen as idempotent by consumers of it. This means if you try to
notify `nginx` to reload on a change, it's always seen as being updated.

This change makes some general enhancements to the `nginx_conf_file` LWRP:

* `use_inline_resources` so that the LWRP can be treated as one atomic unit
* remove the lines always setting `updated_by_last_action` to true
* allow the LWRP to work in `-W/--why-run` mode
* no longer notify `service[nginx]` to restart